// src/types/airport_api.ts
export interface components {
  schemas: {
    Airport: {
      siteNumberCode?: string;
      icao?: string;
      iata?: string | null;
      faaCode?: string;
      city?: string;
      state?: string;
      stateCode?: string;
      country?: string;
      countryCode?: string;
      name?: string;
      keywordPrefixes?: string[];
      artccID?: string;
      artccName?: string;
      artccResponsibleID?: string;
      artccResponsibleName?: string;
      flightServiceStationID?: string;
      flightServiceStationName?: string;
      flightServiceStationPhone?: string;
      flightServiceStationAlternateID?: string | null;
      flightServiceStationAlternateName?: string | null;
      flightServiceStationAlternatePhone?: string | null;
      flightServiceStationIssuingNotamsID?: string | null;
      latitude?: number;
      longitude?: number;
      airportType?: string;
      operationalStatus?: string;
      airportAccess?: string;
      latitudeSecsNorth?: number;
      longitudeSecsEast?: number;
      elevation?: number;
      magneticVariationWest?: number;
      trafficPatternAltitudeMsl?: number | null;
      runways?: components["schemas"]["Runway"][];
      frequencies?: components["schemas"]["Frequency"][];
      remoteFrequencies?: components["schemas"]["Frequency"][];
      procedures?: components["schemas"]["Procedure"][];
      fuelTypes?: string[];
      extraServices?: string[];
      remarks?: components["schemas"]["Remark"][];
      serviceProviders?: string[];
      fbos?: string[];
      businesses?: string[];
      trafficPatternAltitudes?: string[];
      sectionalName?: string;
      managersName?: string | null;
      managersAddress?: string | null;
      managersCityStateZip?: string | null;
      managersPhoneNumber?: string | null;
      afdFilename?: string;
      diagramAvailable?: boolean;
      diagram?: string | null;
      notamDServiceAvailable?: boolean;
      towered?: boolean;
      pdc?: boolean;
      contacts?: components["schemas"]["Contact"][];
      attributes?: components["schemas"]["Attribute"][];
      schedule?: components["schemas"]["Schedule"][];
      cycle?: components["schemas"]["Cycle"];
      timezone?: components["schemas"]["Timezone"];
      source?: string;
      displayName?: string;
      latitudeInDegrees?: number;
      longitudeInDegrees?: number;
    };
    Runway: {
      ident?: string;
      length?: number;
      width?: number;
      surfaceType?: string;
      surfaceQuality?: string;
      edgeLightsIntensity?: string;
      name?: string;
      trueHeading?: number;
      magneticHeading?: number;
      latitudeBase?: number;
      longitudeBase?: number;
      approachSlopeSideBase?: string;
      approachSlopeIndicatorBase?: string;
      approachLightingSystemBase?: string;
      rightTrafficBase?: boolean;
      rightTrafficRecip?: boolean;
      recipTrueHeading?: number;
      recipMagneticHeading?: number;
      recipName?: string;
      latitudeRecip?: number;
      longitudeRecip?: number;
      approachSlopeSideRecip?: string;
      approachSlopeIndicatorRecip?: string;
      approachLightingSystemRecip?: string;
      displacedThresholdBase?: boolean;
      displacedThresholdRecip?: boolean;
      gradientBase?: number;
      gradientRecip?: number;
      displacedThresholdLengthBase?: number | null;
      displacedThresholdLengthRecip?: number | null;
      displacedThresholdLatitudeBase?: number | null;
      displacedThresholdLatitudeRecip?: number | null;
      displacedThresholdLongitudeBase?: number | null;
      displacedThresholdLongitudeRecip?: number | null;
      elevationRunwayEndBase?: number;
      elevationRunwayEndRecip?: number;
      elevationThresholdBase?: number | null;
      elevationThresholdRecip?: number | null;
      elevationTouchdownZoneBase?: number;
      elevationTouchdownZoneRecip?: number;
      glidePathAngleBase?: number | null;
      glidePathAngleRecip?: number | null;
      ldaFtDayBase?: number;
      ldaFtDayRecip?: number;
      ldaFtNightBase?: number;
      ldaFtNightRecip?: number;
      toraFtDayBase?: number;
      toraFtDayRecip?: number;
      toraFtNightBase?: number;
      toraFtNightRecip?: number;
      todaFtDayBase?: number;
      todaFtDayRecip?: number;
      todaFtNightBase?: number;
      todaFtNightRecip?: number;
      asdaFtDayBase?: number;
      asdaFtDayRecip?: number;
      asdaFtNightBase?: number;
      asdaFtNightRecip?: number;
      overrunFtBase?: number;
      overrunFtRecip?: number;
      markingsType?: string | null;
      markingsCondition?: string | null;
      attributes?: Record<string, never> | null;
      edgeLightIntensityKnown?: boolean;
      edgeLightUserPresentable?: string;
      approachSlopeIndicatorBaseAvailable?: boolean;
      approachSlopeIndicatorRecipAvailable?: boolean;
      approachLightingSystemBaseAvailable?: boolean;
      approachLightingSystemRecipAvailable?: boolean;
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}